refactor: Use a persistent PDA for escrowing tokens
The program used to create (and delete) a token account for every
swap. While this makes sense for a data account (swap_account), there
is no real issue with reusing the token account across transactions.
As such, the swap_token_account has been made to be reused for all
transactions involving the same mint.
To authorize the token transfers of this PDA, another PDA named
"identity_pda" has been introduced, which simply represents the
program and does not hold any data (space = 0).