feat: Add field `destination_data` to `initiate()`
Since this program will be participating in an atomic swap that
involves two different chains, this optional field of type `Vec<u8>` shall
store additional information regarding the target chain.

The contract doesn't act on this data, it merely emits this in its
events, allowing for listeners to have context directly from the chain,
rather than from an off-chain source.