feat: Separate initiator into `funder` and `refundee`
The initiator in an atomic swap has two roles, one of depositing funds
and the other of obtaining refunds.
For cases where these two need to be different, e.g an initiator
wants to participate but cannot provide their signature,
they can do so through a party referred to as the "funder" who
deposits funds on the initiator's behalf.
And if the swap fails, the initiator aka refundee gets back the funds.