fix: Verify mint for all token accounts
The current checks for redeem/refund instructions only verify:
- Does the token vault belong to the program?
- Does the provided token accounts belong to the right people

The lack of checking mint allows for a malicious user to provide a valid token vault and a valid token
account for a mint that they did not initiate with.

Example scenario:
1. User A initiates with Mint M1 and User B as redeemer.
2. A's funds of mint M1 are stored in token vault T1.
3. User C initiates with Mint M2 and User D as redeemer.
4. C's funds of mint M2 are stored in token vault T2.
5. D "redeems" by specifying
  - token vault T1 instead of T2
  - D's token account of mint M1 instead of M2

Effectively stealing A's funds from token vault T1.

Thus store mint in the swap account and use it to verify the mints
of all the token accounts.